Implement auth logic
Created the static class ApplicationState with the boolean member isAdmin and implemented the logic in the Login button and Dictionary class constructor.

Now when trying to log in as an admin, your credentials get compared to a JSON file of "admins". If your credentials match up with one of the objects from the JSON the ApplicationState updates it's isAdmin member to true.

Now depending on your choice of login, the state of the application knows if you're a normal user or an admin.

I also added in the top-right corner of the Dictionary page a TextBox to notify that admin mode is on when you log in as an administrator.