Read Categories and Words from JSON and store them
Created the classes WordList and CategoryMap that are used for deserializing the JSON files "Words.json" and "Categories.json". Both of the classes have a function
readData that are used to initialize the variables "words" and "categories" from the ApplicationState class.

Now at the the start of the app, the ApplicationState class has available all the words and categories from the JSONs.